Add Kafka publishing for patches
When EDIT Update type messages are received by the backend and have been
processed, a goroutine is started that will publish the processed
message to a Kafka topic. This topic should be read by ether in order to
update the content file and eventually should also be read by redundant
copies of patches that also have client connections for the same
conversation.