Update tf to fix Heimdall internal server access
It needs its own cluster to so that it is running on a separate EC2
instance from everything else. It needs to do this because of the way
Network Load Balancers work: they don't like requests looping back to
the same instance.

Also updated the Heimdall module config to use the new variables.